:test: ajusta erros em teste do feed.js
