+++ src/firebase/__mocks__/auth.js
@@ -1,3 +1,5 @@
 export const loginUser = jest.fn(() => Promise.resolve({}));
 export const loginGoogle = jest.fn(() => Promise.resolve({}));
 export const newUser = jest.fn(() => Promise.resolve({}));
+export const logout = jest.fn(() => Promise.resolve({}));
+
